adding bad scalar value to gToMinutes

+++ src/main/java/model/Path.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 public class Path {
 
     public ArrayList<Node> nodes;
-    public int time;
+    public double time;
 
-    public Path(ArrayList<Node> nodes, int time) {
+    public Path(ArrayList<Node> nodes, double time) {
         this.nodes = nodes;
         this.time = time;
     }
